Added workflow to build and publish client library (#46)
Fixed #41

I have added a GitHub workflow that when a semver tag is pushed to git, then a workflow to build and publish the nuget packages is created. It also creates a GitHub release, where one can see the changes made, it supports two formats for know:

- v1.2.3
- v1.2.3-alpha.4

when a tag with the alpha suffix is created, it creates the GitHub release as a PreRelease.