Update roadmap.md
"It will be all in one container" what does "it" refer to? 

I don't know what the sentence "proving the federated principle with in process calls." means.

Why are the "Long term goals" not included in the alpha and beta versions, or why is it not a third (gamma) version?