Update full.md
l. 8: A little bit of context is needed here. Is a consumption certificate created first, requesting a production certificate of the consumption amount? And then the production amount is sliced? Or how is the claim transaction initiated?

l. 69: mentions that the remainer should be saved, but it isn't in the code example - maybe just move the codeline down, so that it is not part of the comment